2026.7.13

+++ python/2026.7.10/P1089.py
@@ -12,6 +12,6 @@ def main():
             cun+=100*((shou-yu)//100)
             shou-=100*((shou-yu)//100)
         shou-=yu
-    print(shou+cun*1.2)
+    print(int(shou+cun*1.2))
 
 main()
